backend/requirements.txt
# Core backend
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9

# ML / Data
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2

# Deep learning (disease detection)
torch==2.3.0
torchvision==0.18.0
Pillow==10.3.0

# Explainability
shap==0.45.1
matplotlib==3.9.0

# Weather API
httpx==0.27.0

# Utilities
python-dotenv==1.0.1
pydantic==2.7.1
